requirements.txt
alembic==1.5.8
aniso8601==9.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
mysql-connector-python==8.0.24
packaging==20.9
pluggy==0.13.1
protobuf==3.15.8
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.10
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
